🚧 Rename class to CartTransformer
